gitcohealth/backend/requirements.txt
Flask
requests
python-dotenv
openai
flask-cors
